fix(link-adaptation): restore exact legacy PHY arithmetic after Round 12 regression
- restore pre-v0.13 float32 CQI-efficiency calculations in legacy mode
- prevent tiny numerical reward drift from changing PPO training trajectories
- add regression coverage for legacy attempted-bit calculations
- leave CQI-to-MCS and channel-dependent BLER behavior unchanged

+++ scalemac_rl/env.py
@@ -597,7 +597,7 @@ class ScaleMacDownlinkEnv:
 
         if selected.size:
             if self.config.link_adaptation_mode == "legacy_fixed_bler":
-                efficiency = _CQI_EFFICIENCY[self.cqi[selected] - 1].astype(np.float64)
+                efficiency = _CQI_EFFICIENCY[self.cqi[selected] - 1]
                 predicted_bler.fill(self.config.target_bler if self.config.harq_enabled else 0.0)
                 selected_mcs.fill(-1)
                 selected_mod_order.fill(0)
